requirements.txt
docker-py==1.10.6
docker-pycreds==0.2.1
python-consul==0.7.0
requests==2.12.4
six==1.10.0
websocket-client==0.40.0
